refactor(commands): move resumeArchivedDownloads to commands/resume.cjs
